[Feat] FreeBoardWirte & FreeBoard 페이지 가생성 및 라우터 설정
