+++ src/pages/FreeBoardWrite.module.css
@@ -1,0 +1,44 @@
+.registrationPage {
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+}
+
+nav {
+  width: 100%;
+}
+
+.mainContainer {
+  display: flex;
+  flex-direction: column;
+  width: 100%;
+  max-width: 1200px;
+  padding: 40px 24px;
+  flex-basis: 1000px;
+  gap: 16px 0px
+}
+
+.inputSection {
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  gap: 24px;
+}
+
+
+.registrationHeader {
+  width: 100%;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  gap: 12px;
+}
+
+.headerText {
+  font-weight: 700;
+  font-size: 1.25rem;
+}
